+++ crates/rust-client/src/account/mod.rs
@@ -140,11 +140,18 @@ impl<AUTH> Client<AUTH> {
     ///   being tracked.
     /// - If `overwrite` is set to `true` and the `account_data` commitment doesn't match the
     ///   network's account commitment.
+    /// - If the client has reached the accounts limit
+    ///   ([`ACCOUNT_ID_LIMIT`](crate::rpc::ACCOUNT_ID_LIMIT)).
+    /// - If the client has reached the note tags limit
+    ///   ([`NOTE_TAG_LIMIT`](crate::rpc::NOTE_TAG_LIMIT)).
     pub async fn add_account(
         &mut self,
         account: &Account,
         overwrite: bool,
     ) -> Result<(), ClientError> {
+        self.check_account_limit().await?;
+        self.check_note_tag_limit().await?;
+
         if account.is_new() {
             if account.seed().is_none() {
                 return Err(ClientError::AddNewAccountWithoutSeed);
@@ -234,6 +241,8 @@ impl<AUTH> Client<AUTH> {
     /// # Errors
     /// - If the account is not found on the network.
     /// - If the address is already being tracked.
+    /// - If the client has reached the note tags limit
+    ///   ([`NOTE_TAG_LIMIT`](crate::rpc::NOTE_TAG_LIMIT)).
     pub async fn add_address(
         &mut self,
         address: Address,
@@ -260,6 +269,7 @@ impl<AUTH> Client<AUTH> {
                     ));
                 }
 
+                self.check_note_tag_limit().await?;
                 self.store.insert_address(address, account_id).await?;
                 Ok(())
             },